# backend/src/routes/wallets.ts
import { Router } from "express";
import { openai } from "@workspace/integrations-openai-ai-server";
import { GetWalletDataParams, AnalyzeWalletParams } from "@workspace/api-zod";
import { walletLimiter, aiLimiter } from "../lib/rateLimit";
import { createPublicClient, http } from "viem";
import { mainnet } from "viem/chains";

const router = Router();

const ETHERSCAN_API = "https://api.etherscan.io/v2/api";
const ETH_PRICE_API = "https://api.coingecko.com/api/v3/simple/price?ids=ethereum&vs_currencies=usd";
const ETHERSCAN_KEY = process.env["ETHERSCAN_API_KEY"] ?? "";

const publicClient = createPublicClient({
  chain: mainnet,
  transport: http(process.env["ETH_RPC_URL"] ?? "https://eth.drpc.org"),
});

async function ensToAddress(ensName: string): Promise<string | null> {
  try {
    const normalized = ensName.toLowerCase().trim() as `${string}.eth`;
    const address = await publicClient.getEnsAddress({ name: normalized });
    return address ?? null;
  } catch {
    return null;
  }
}

function etherscanUrl(params: Record<string, string>): string {
  const p = new URLSearchParams({ chainid: "1", ...params });
  if (ETHERSCAN_KEY) p.set("apikey", ETHERSCAN_KEY);
  return `${ETHERSCAN_API}?${p.toString()}`;
}

async function getEthPrice(): Promise<number> {
  try {
    const res = await fetch(ETH_PRICE_API);
    const data = await res.json() as { ethereum?: { usd?: number } };
    return data?.ethereum?.usd ?? 0;
  } catch {
    return 0;
  }
}

async function resolveEnsName(address: string): Promise<string | null> {
  try {
    const url = `https://api.ensideas.com/ens/resolve/${address}`;
    const res = await fetch(url);
    if (!res.ok) return null;
    const data = await res.json() as { name?: string };
    return data?.name ?? null;
  } catch {
    return null;
  }
}

interface WalletSnapshot {
  balance: string;
  txCount: number;
  lastTxs: Awaited<ReturnType<typeof getLastTransactions>>;
  ensName: string | null;
  fetchedAt: number;
}

const walletCache = new Map<string, WalletSnapshot>();
const WALLET_CACHE_TTL = 5 * 60 * 1000;

const pendingSnapshots = new Map<string, Promise<WalletSnapshot>>();

async function fetchWalletSnapshot(address: string, knownEns: string | null): Promise<WalletSnapshot> {
  const key = address.toLowerCase();
  const cached = walletCache.get(key);
  if (cached && Date.now() - cached.fetchedAt < WALLET_CACHE_TTL) return cached;

  const existing = pendingSnapshots.get(key);
  if (existing) return existing;

  const promise = (async (): Promise<WalletSnapshot> => {
    const doFetch = async (): Promise<WalletSnapshot> => {
      const [balance, txCount, lastTxs, ensName] = await Promise.all([
        getWalletBalance(address),
        getTransactionCount(address),
        getLastTransactions(address),
        knownEns ? Promise.resolve(knownEns) : resolveEnsName(address),
      ]);
      return { balance, txCount, lastTxs, ensName, fetchedAt: Date.now() };
    };

    let snapshot: WalletSnapshot;
    try {
      snapshot = await doFetch();
    } catch (err) {
      if (err instanceof Error && err.message === "etherscan_rate_limit") {
        await new Promise((r) => setTimeout(r, 2500));
        snapshot = await doFetch();
      } else {
        throw err;
      }
    }
    walletCache.set(key, snapshot);
    pendingSnapshots.delete(key);
    return snapshot;
  })();

  pendingSnapshots.set(key, promise);
  return promise;
}

function isRateLimitResult(result: string | undefined): boolean {
  if (!result) return false;
  const r = result.toLowerCase();
  return r.includes("rate limit") || r.includes("max rate") || r.includes("notok");
}

async function getWalletBalance(address: string): Promise<string> {
  const url = etherscanUrl({ module: "account", action: "balance", address, tag: "latest" });
  const res = await fetch(url);
  const data = await res.json() as { result?: string; status?: string };
  if (isRateLimitResult(data?.result) || (data?.status === "0" && isRateLimitResult(data?.result))) {
    throw new Error("etherscan_rate_limit");
  }
  try {
    const weiBalance = BigInt(data?.result ?? "0");
    const ethBalance = Number(weiBalance) / 1e18;
    return ethBalance.toFixed(4);
  } catch {
    return "0";
  }
}

async function getTransactionCount(address: string): Promise<number> {
  const url = etherscanUrl({ module: "proxy", action: "eth_getTransactionCount", address, tag: "latest" });
  const res = await fetch(url);
  const data = await res.json() as { result?: string };
  if (isRateLimitResult(data?.result)) throw new Error("etherscan_rate_limit");
  const n = parseInt(data?.result ?? "0x0", 16);
  return isNaN(n) ? 0 : n;
}

async function getLastTransactions(address: string): Promise<Array<{
  hash: string;
  from: string;
  to: string | null;
  value: string;
  valueEth: string;
  gasUsed: string;
  timestamp: number;
  method: string | null;
  status: string;
}>> {
  try {
    const url = etherscanUrl({ module: "account", action: "txlist", address, startblock: "0", endblock: "99999999", page: "1", offset: "3", sort: "desc" });
    const res = await fetch(url);
    const data = await res.json() as {
      status?: string;
      result?: Array<{
        hash: string;
        from: string;
        to: string;
        value: string;
        gasUsed: string;
        timeStamp: string;
        functionName: string;
        txreceipt_status: string;
        isError: string;
      }>;
    };

    if (data?.status !== "1" || !Array.isArray(data.result)) return [];

    return data.result.slice(0, 3).map((tx) => {
      const valueEth = (Number(tx.value) / 1e18).toFixed(4);
      const method = tx.functionName
        ? tx.functionName.split("(")[0]
        : tx.value !== "0"
        ? "Transfer"
        : null;

      return {
        hash: tx.hash,
        from: tx.from,
        to: tx.to || null,
        value: tx.value,
        valueEth,
        gasUsed: tx.gasUsed,
        timestamp: parseInt(tx.timeStamp),
        method,
        status: tx.txreceipt_status === "1" || tx.isError === "0" ? "success" : "failed",
      };
    });
  } catch {
    return [];
  }
}

router.get("/wallets/:address", walletLimiter, async (req, res) => {
  const parsed = GetWalletDataParams.safeParse(req.params);
  if (!parsed.success) {
    res.status(400).json({ error: "Invalid address" });
    return;
  }

  const { address: rawAddress } = parsed.data;

  if (!/^0x[0-9a-fA-F]{40}$/.test(rawAddress) && !rawAddress.endsWith(".eth")) {
    res.status(400).json({ error: "Invalid Ethereum address or ENS name" });
    return;
  }

  try {
    let address = rawAddress;
    let inputEnsName: string | null = null;

    if (rawAddress.endsWith(".eth")) {
      inputEnsName = rawAddress;
      const resolved = await ensToAddress(rawAddress);
      if (!resolved) {
        res.status(404).json({ error: `ENS name ${rawAddress} could not be resolved to an address` });
        return;
      }
      address = resolved;
    }

    const [snap, ethPrice] = await Promise.all([
      fetchWalletSnapshot(address, inputEnsName),
      getEthPrice(),
    ]);

    const balanceUsd = ethPrice > 0
      ? (parseFloat(snap.balance) * ethPrice).toFixed(2)
      : null;

    res.json({
      address,
      ensName: snap.ensName,
      balanceEth: snap.balance,
      balanceUsd,
      txCount: snap.txCount,
      lastTransactions: snap.lastTxs,
      firstSeen: snap.lastTxs.length > 0
        ? new Date(Math.min(...snap.lastTxs.map((t) => t.timestamp * 1000))).toISOString()
        : null,
      isContract: false,
    });
  } catch (err) {
    req.log.error({ err }, "Failed to get wallet data");
    res.status(500).json({ error: "Internal server error" });
  }
});

router.get("/wallets/:address/analyze", walletLimiter, aiLimiter, async (req, res) => {
  const parsed = AnalyzeWalletParams.safeParse(req.params);
  if (!parsed.success) {
    res.status(400).json({ error: "Invalid address" });
    return;
  }

  const { address: rawAddress2 } = parsed.data;

  try {
    let address = rawAddress2;
    let inputEnsName2: string | null = null;

    if (rawAddress2.endsWith(".eth")) {
      inputEnsName2 = rawAddress2;
      const resolved = await ensToAddress(rawAddress2);
      if (!resolved) {
        res.status(404).json({ error: `ENS name ${rawAddress2} could not be resolved to an address` });
        return;
      }
      address = resolved;
    }

    const snap = await fetchWalletSnapshot(address, inputEnsName2);
    const { balance, txCount, lastTxs, ensName } = { balance: snap.balance, txCount: snap.txCount, lastTxs: snap.lastTxs, ensName: snap.ensName };

    const txSummary = lastTxs
      .map(
        (tx) =>
          `- Hash: ${tx.hash.slice(0, 10)}... | From: ${tx.from.slice(0, 8)}... | To: ${tx.to?.slice(0, 8) ?? "contract"} | ${tx.valueEth} ETH | Method: ${tx.method ?? "unknown"} | Status: ${tx.status} | Time: ${new Date(tx.timestamp * 1000).toISOString()}`
      )
      .join("\n");

    const prompt = `You are a Web3 on-chain analyst. Analyze this Ethereum wallet and provide concise insights.

Wallet: ${address}
ENS Name: ${ensName ?? "none"}
Balance: ${balance} ETH
Total TX Count: ${txCount}
Recent Transactions:
${txSummary || "No recent transactions found"}

Provide a JSON response with:
{
  "summary": "2-3 sentence overview of this wallet's behavior",
  "activityType": "one of: DeFi Trader, NFT Collector, Long-term Holder, Power User, Exchange Wallet, Inactive Wallet, Active Transactor",
  "riskLevel": "low | medium | high",
  "tags": ["array", "of", "3-5", "descriptive", "tags"],
  "insights": ["array of 3-4 specific insights about this wallet's on-chain activity"]
}

Respond ONLY with valid JSON, no markdown. Never use em dashes or double hyphens in any text values.`;

    const completion = await openai.chat.completions.create({
      model: "gpt-5.2",
      max_completion_tokens: 1024,
      messages: [{ role: "user", content: prompt }],
    });

    const raw = completion.choices[0]?.message?.content ?? "{}";
    let analysis: {
      summary: string;
      activityType: string;
      riskLevel: string;
      tags: string[];
      insights: string[];
    };

    try {
      analysis = JSON.parse(raw);
    } catch {
      analysis = {
        summary: raw,
        activityType: "Unknown",
        riskLevel: "low",
        tags: [],
        insights: [],
      };
    }

    const BANNED_ACTIVITY_TYPES = ["developer"];
    const rawActivityType = analysis.activityType ?? "Unknown";
    const sanitizedActivityType = BANNED_ACTIVITY_TYPES.includes(rawActivityType.toLowerCase())
      ? "Power User"
      : rawActivityType;

    res.json({
      address,
      summary: analysis.summary ?? "",
      activityType: sanitizedActivityType,
      riskLevel: (["low", "medium", "high"].includes(analysis.riskLevel) ? analysis.riskLevel : "low") as "low" | "medium" | "high",
      tags: Array.isArray(analysis.tags) ? analysis.tags : [],
      insights: Array.isArray(analysis.insights) ? analysis.insights : [],
      generatedAt: new Date().toISOString(),
    });
  } catch (err) {
    req.log.error({ err }, "Failed to analyze wallet");
    res.status(500).json({ error: "Internal server error" });
  }
});

export default router;