# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Subframe Protocol API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallets/{address}:
    get:
      operationId: getWalletData
      tags: [wallets]
      summary: Get wallet on-chain data including last 3 transactions
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Wallet data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletData"
        "400":
          description: Invalid address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /wallets/{address}/analyze:
    get:
      operationId: analyzeWallet
      tags: [wallets]
      summary: Get AI analysis of a wallet's on-chain activity
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletAnalysis"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Transaction:
      type: object
      properties:
        hash:
          type: string
        from:
          type: string
        to:
          type: ["string", "null"]
        value:
          type: string
        valueEth:
          type: string
        gasUsed:
          type: string
        timestamp:
          type: integer
        method:
          type: ["string", "null"]
        status:
          type: string
      required:
        - hash
        - from
        - to
        - value
        - valueEth
        - gasUsed
        - timestamp
        - method
        - status
    WalletData:
      type: object
      properties:
        address:
          type: string
        ensName:
          type: ["string", "null"]
        balanceEth:
          type: string
        balanceUsd:
          type: ["string", "null"]
        txCount:
          type: integer
        lastTransactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        firstSeen:
          type: ["string", "null"]
        isContract:
          type: boolean
      required:
        - address
        - ensName
        - balanceEth
        - balanceUsd
        - txCount
        - lastTransactions
        - firstSeen
        - isContract
    WalletAnalysis:
      type: object
      properties:
        address:
          type: string
        summary:
          type: string
        activityType:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        tags:
          type: array
          items:
            type: string
        insights:
          type: array
          items:
            type: string
        generatedAt:
          type: string
          format: date-time
      required:
        - address
        - summary
        - activityType
        - riskLevel
        - tags
        - insights
        - generatedAt
...